+++ __tests__/components/Header/Header.test.js
@@ -12,12 +12,14 @@ const defaultProps = {
   minUnit: 'day',
   timeSteps: {},
   width: 400,
-  headerLabelFormats: {},
-  subHeaderLabelFormats: {},
+  topHeaderLabelFormats: {},
+  middleHeaderLabelFormats: {},
+  bottomHeaderLabelFormats: {},
   stickyOffset: 5,
   stickyHeader: true,
-  headerLabelGroupHeight: 15,
-  headerLabelHeight: 15,
+  topHeaderLabelHeight: 15,
+  middleHeaderLabelHeight: 15,
+  bottomHeaderLabelHeight: 15,
   registerScroll: () => {},
   headerRef: () => {}
 }