+++ src/lib/layout/TimelineElementsHeader.js
@@ -14,10 +14,13 @@ export default class TimelineElementsHeader extends Component {
     minUnit: PropTypes.string.isRequired,
     timeSteps: PropTypes.object.isRequired,
     width: PropTypes.number.isRequired,
-    headerLabelFormats: PropTypes.object.isRequired,
-    subHeaderLabelFormats: PropTypes.object.isRequired,
-    headerLabelGroupHeight: PropTypes.number.isRequired,
-    headerLabelHeight: PropTypes.number.isRequired,
+    topHeaderLabelFormats: PropTypes.object.isRequired,
+    middleHeaderLabelFormats: PropTypes.object.isRequired,
+    bottomHeaderLabelFormats: PropTypes.object.isRequired,
+    topHeaderLabelHeight: PropTypes.number.isRequired,
+    middleHeaderLabelHeight: PropTypes.number.isRequired,
+    bottomHeaderLabelHeight: PropTypes.number.isRequired,
+    useThreeRowHeader: PropTypes.bool.isRequired,
     registerScroll: PropTypes.func.isRequired
   }
 
@@ -41,8 +44,8 @@ export default class TimelineElementsHeader extends Component {
     evt.stopPropagation()
   }
 
-  headerLabel(time, unit, width) {
-    const { headerLabelFormats: f } = this.props
+  topHeaderLabel(time, unit, width) {
+    const { topHeaderLabelFormats: f } = this.props
 
     if (unit === 'year') {
       return time.format(width < 46 ? f.yearShort : f.yearLong)
@@ -54,6 +57,8 @@ export default class TimelineElementsHeader extends Component {
             ? f.monthMedium
             : width < 120 ? f.monthMediumLong : f.monthLong
       )
+    } else if (unit === 'week') {
+      return time.format(width < 46 ? f.weekMedium : f.weekLong)
     } else if (unit === 'day') {
       return time.format(width < 150 ? f.dayShort : f.dayLong)
     } else if (unit === 'hour') {
@@ -69,8 +74,8 @@ export default class TimelineElementsHeader extends Component {
     }
   }
 
-  subHeaderLabel(time, unit, width) {
-    const { subHeaderLabelFormats: f } = this.props
+  middleHeaderLabel(time, unit, width) {
+    const { middleHeaderLabelFormats: f } = this.props
 
     if (unit === 'year') {
       return time.format(width < 46 ? f.yearShort : f.yearLong)
@@ -78,6 +83,28 @@ export default class TimelineElementsHeader extends Component {
       return time.format(
         width < 37 ? f.monthShort : width < 85 ? f.monthMedium : f.monthLong
       )
+    } else if (unit === 'week') {
+      return time.format(f.weekMedium)
+    } else if (unit === 'day') {
+      return time.format(width < 47 ? f.dayShort : f.dayLong)
+    } else if (unit === 'hour') {
+      return time.format(f.hourLong);
+    } else {
+      return time.format(f.time)
+    }
+  }
+
+  bottomHeaderLabel(time, unit, width) {
+    const { bottomHeaderLabelFormats: f } = this.props
+
+    if (unit === 'year') {
+      return time.format(width < 46 ? f.yearShort : f.yearLong)
+    } else if (unit === 'month') {
+      return time.format(
+        width < 37 ? f.monthShort : width < 85 ? f.monthMedium : f.monthLong
+      )
+    } else if (unit === 'week') {
+      return time.format(width < 46 ? f.weekShort : f.weekMedium)
     } else if (unit === 'day') {
       return time.format(
         width < 47
@@ -116,18 +143,21 @@ export default class TimelineElementsHeader extends Component {
       canvasWidth,
       minUnit,
       timeSteps,
-      headerLabelGroupHeight,
-      headerLabelHeight,
+      topHeaderLabelHeight,
+      middleHeaderLabelHeight,
+      bottomHeaderLabelHeight,
+      useThreeRowHeader,
       hasRightSidebar
     } = this.props
 
     const ratio = canvasWidth / (canvasTimeEnd - canvasTimeStart)
     const twoHeaders = minUnit !== 'year'
+    const threeHeaders = useThreeRowHeader && minUnit !== 'month' && twoHeaders
 
     const topHeaderLabels = []
     // add the top header
-    if (twoHeaders) {
-      const nextUnit = getNextUnit(minUnit)
+    if (threeHeaders) {
+      const nextUnit = getNextUnit(getNextUnit(minUnit))
 
       iterateTimes(
         canvasTimeStart,
@@ -150,20 +180,66 @@ export default class TimelineElementsHeader extends Component {
           topHeaderLabels.push(
             <div
               key={`top-label-${time.valueOf()}`}
-              className={`rct-label-group${
+              className={`rct-label-top${
                 hasRightSidebar ? ' rct-has-right-sidebar' : ''
               }`}
               onClick={() => this.handlePeriodClick(time, nextUnit)}
               style={{
                 left: `${left - 1}px`,
                 width: `${labelWidth}px`,
-                height: `${headerLabelGroupHeight}px`,
-                lineHeight: `${headerLabelGroupHeight}px`,
+                height: `${topHeaderLabelHeight}px`,
+                lineHeight: `${topHeaderLabelHeight}px`,
                 cursor: 'pointer'
               }}
             >
               <span style={{ width: contentWidth, display: 'block' }}>
-                {this.headerLabel(time, nextUnit, labelWidth)}
+                {this.topHeaderLabel(time, nextUnit, labelWidth)}
+              </span>
+            </div>
+          )
+        }
+      )
+    }
+
+    const middleHeaderLabels = []
+    if (twoHeaders) {
+      const nextUnit = getNextUnit(minUnit)
+
+      iterateTimes(
+        canvasTimeStart,
+        canvasTimeEnd,
+        nextUnit,
+        timeSteps,
+        (time, nextTime) => {
+          const left = Math.round((time.valueOf() - canvasTimeStart) * ratio)
+          const right = Math.round(
+            (nextTime.valueOf() - canvasTimeStart) * ratio
+          )
+
+          const labelWidth = right - left
+          // this width applies to the content in the header
+          // it simulates stickyness where the content is fixed in the center
+          // of the label.  when the labelWidth is less than visible time range,
+          // have label content fill the entire width
+          const contentWidth = Math.min(labelWidth, canvasWidth / 3)
+
+          middleHeaderLabels.push(
+            <div
+              key={`middle-label-${time.valueOf()}`}
+              className={`rct-label-middle${
+                hasRightSidebar ? ' rct-has-right-sidebar' : ''
+              }`}
+              onClick={() => this.handlePeriodClick(time, nextUnit)}
+              style={{
+                left: `${left - 1}px`,
+                width: `${labelWidth}px`,
+                height: `${middleHeaderLabelHeight}px`,
+                lineHeight: `${middleHeaderLabelHeight}px`,
+                cursor: 'pointer'
+              }}
+            >
+              <span style={{width: contentWidth, display: 'block'}}>
+                {this.middleHeaderLabel(time, nextUnit, labelWidth)}
               </span>
             </div>
           )
@@ -189,7 +265,7 @@ export default class TimelineElementsHeader extends Component {
         bottomHeaderLabels.push(
           <div
             key={`label-${time.valueOf()}`}
-            className={`rct-label ${twoHeaders ? '' : 'rct-label-only'} ${
+            className={`rct-label-bottom ${twoHeaders ? '' : 'rct-label-only'} ${
               firstOfType ? 'rct-first-of-type' : ''
             } ${minUnit !== 'month' ? `rct-day-${time.day()}` : ''} `}
             onClick={() => this.handlePeriodClick(time, minUnit)}
@@ -197,14 +273,14 @@ export default class TimelineElementsHeader extends Component {
               left: `${left - leftCorrect}px`,
               width: `${labelWidth}px`,
               height: `${
-                minUnit === 'year'
-                  ? headerLabelGroupHeight + headerLabelHeight
-                  : headerLabelHeight
+                threeHeaders ? bottomHeaderLabelHeight
+                  : twoHeaders ? (useThreeRowHeader ? bottomHeaderLabelHeight + topHeaderLabelHeight : bottomHeaderLabelHeight)
+                  :(useThreeRowHeader ? bottomHeaderLabelHeight + middleHeaderLabelHeight + topHeaderLabelHeight : bottomHeaderLabelHeight + middleHeaderLabelHeight)
               }px`,
               lineHeight: `${
-                minUnit === 'year'
-                  ? headerLabelGroupHeight + headerLabelHeight
-                  : headerLabelHeight
+                threeHeaders ? bottomHeaderLabelHeight
+                  : twoHeaders ? (useThreeRowHeader ? bottomHeaderLabelHeight + topHeaderLabelHeight : bottomHeaderLabelHeight)
+                  :(useThreeRowHeader ? bottomHeaderLabelHeight + middleHeaderLabelHeight + topHeaderLabelHeight : bottomHeaderLabelHeight + middleHeaderLabelHeight)
               }px`,
               fontSize: `${
                 labelWidth > 30 ? '14' : labelWidth > 20 ? '12' : '10'
@@ -212,14 +288,14 @@ export default class TimelineElementsHeader extends Component {
               cursor: 'pointer'
             }}
           >
-            {this.subHeaderLabel(time, minUnit, labelWidth)}
+            {this.bottomHeaderLabel(time, minUnit, labelWidth)}
           </div>
         )
       }
     )
 
     let headerStyle = {
-      height: `${headerLabelGroupHeight + headerLabelHeight}px`
+      height: `${useThreeRowHeader ? topHeaderLabelHeight + middleHeaderLabelHeight + bottomHeaderLabelHeight : middleHeaderLabelHeight + bottomHeaderLabelHeight}px`
     }
 
     return (
@@ -235,13 +311,24 @@ export default class TimelineElementsHeader extends Component {
       >
         <div
           className="top-header"
-          style={{ height: twoHeaders ? headerLabelGroupHeight : 0, width: canvasWidth }}
+          style={{ height: threeHeaders ? topHeaderLabelHeight : 0, width: canvasWidth }}
         >
           {topHeaderLabels}
         </div>
         <div
+          className="middle-header"
+          style={{ height: twoHeaders ? middleHeaderLabelHeight : 0, width: canvasWidth}}>
+          {middleHeaderLabels}
+        </div>
+        <div
           className="bottom-header"
-          style={{ height: twoHeaders ? headerLabelHeight : headerLabelHeight + headerLabelGroupHeight, width: canvasWidth }}
+          style={{
+            height: threeHeaders
+              ? bottomHeaderLabelHeight : twoHeaders
+              ? (useThreeRowHeader ? bottomHeaderLabelHeight + topHeaderLabelHeight : bottomHeaderLabelHeight)
+              : (useThreeRowHeader ? bottomHeaderLabelHeight + middleHeaderLabelHeight + topHeaderLabelHeight : bottomHeaderLabelHeight + middleHeaderLabelHeight),
+            width: canvasWidth
+          }}
         >
           {bottomHeaderLabels}
         </div>